requirements.txt
aerich==0.7.2
aiogram==2.21
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dictdiffer==0.9.0
flake8==6.1.0
frozenlist==1.4.0
idna==3.4
iso8601==1.1.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
pypika-tortoise==0.1.6
python-dotenv==0.20.0
pytz==2023.3.post1
tomli==2.0.1
tomlkit==0.12.1
tortoise-orm==0.20.0
typing_extensions==4.8.0
yarl==1.9.2
